Refactor from npm workspaces to simple setup
