+++ .github/workflows/ci.yml
@@ -5,7 +5,7 @@ on:
   pull_request:
 
 jobs:
-  build-api:
+  build:
     runs-on: ubuntu-latest
     services:
       postgres:
@@ -23,7 +23,6 @@ jobs:
           --health-retries 5
     env:
       DATABASE_URL: postgres://postgres:postgres@localhost:5432/simple_subnet_api
-      NPM_CONFIG_WORKSPACE: api
     steps:
       - uses: actions/checkout@v4
       - uses: actions/setup-node@v4
@@ -32,17 +31,6 @@ jobs:
       - run: npm ci
       - run: npm run migrate
       - run: npm test
-
-  lint-all:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-      - uses: actions/setup-node@v4
-        with:
-          node-version: 22
-      - run: npm ci
-      - run: npm run lint
-      - run: npm run test:types
 
   docker-build:
     runs-on: ubuntu-latest
@@ -69,12 +57,12 @@ jobs:
 
   deploy-api:
     if: github.ref == 'refs/heads/main'
-    needs: [build-api, build-backend, docker-build]
+    needs: [build, docker-build]
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
       - uses: superfly/flyctl-actions/setup-flyctl@master
-      - run: flyctl deploy --remote-only -c api/fly.toml
+      - run: flyctl deploy --remote-only -c fly.toml
         env:
           FLY_API_TOKEN: ${{ secrets.FLY_TOKEN_API }}
       - if: failure()
