--- api/lib/typings.d.ts
@@ -1,5 +1,0 @@
-export interface Logger {
-  info: typeof console.info;
-  error: typeof console.error;
-  request: typeof console.info;
-}
