--- api/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "@filecoin-station/simple-subnet-api",
-  "type": "module",
-  "private": true,
-  "scripts": {
-    "migrate": "node ../db/bin/migrate.js",
-    "start": "node bin/simple-subnet-api.js",
-    "lint": "standard",
-    "test": "node --test --test-reporter=spec"
-  },
-  "devDependencies": {
-    "standard": "^17.1.2"
-  },
-  "dependencies": {
-    "@fastify/postgres": "^6.0.2",
-    "@filecoin-station/simple-subnet-api-db": "^1.0.0",
-    "@sentry/node": "^9.2.0",
-    "@sentry/profiling-node": "^9.2.0",
-    "debug": "^4.4.0",
-    "fastify": "^5.2.1",
-    "http-assert": "^1.5.0",
-    "pg": "^8.13.3"
-  }
-}
