--- db/bin/migrate.js
@@ -1,4 +1,0 @@
-import { migrateWithPgClient, createPgPool } from '../index.js'
-
-const pgPool = await createPgPool()
-await migrateWithPgClient(pgPool)
