--- db/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "@filecoin-station/simple-subnet-api-db",
-  "version": "1.0.0",
-  "type": "module",
-  "main": "index.js",
-  "private": true,
-  "scripts": {
-    "migrate": "node bin/migrate.js",
-    "lint": "standard"
-  },
-  "devDependencies": {
-    "standard": "^17.1.2"
-  },
-  "dependencies": {
-    "pg": "^8.13.3",
-    "postgrator": "^8.0.0"
-  }
-}
