+++ fly.toml
@@ -9,7 +9,6 @@ PORT = "8080"
 HOST = "0.0.0.0"
 SENTRY_ENVIRONMENT = "production"
 REQUEST_LOGGING = "false"
-NPM_CONFIG_WORKSPACE = "api"
 
 [experimental]
 auto_rollback = true
