+++ lib/instrument.js
