+++ lib/migrate.js
@@ -12,11 +12,6 @@ import Postgrator from 'postgrator'
 // Configure node-postgres to deserialize BIGINT values as BigInt, not String
 pg.types.setTypeParser(20, BigInt) // Type Id 20 = BIGINT | BIGSERIAL
 
-const {
-  // DATABASE_URL points to `spark_deal_observer` database managed by this monorepo
-  DATABASE_URL = 'postgres://localhost:5432/simple_subnet_api'
-} = process.env
-
 const migrationsDirectory = join(
   dirname(fileURLToPath(import.meta.url)),
   'migrations'
@@ -47,12 +42,13 @@ const onError = (err) => {
 }
 
 /**
+ * @param {string} connectionString
  * @returns {Promise<PgPool>}
  */
-export const createPgPool = async () => {
+export const createPgPool = async (connectionString) => {
   const pool = new pg.Pool({
     ...poolConfig,
-    connectionString: DATABASE_URL
+    connectionString 
   })
   pool.on('error', onError)
   await pool.query('SELECT 1')
@@ -81,5 +77,3 @@ export const migrateWithPgClient = async (client) => {
     await postgrator.getDatabaseVersion()
   )
 }
-
-export { DATABASE_URL }
