+++ lib/typings.d.ts
@@ -1,5 +1,11 @@
 import pg from "pg";
 
+export interface Logger {
+  info: typeof console.info;
+  error: typeof console.error;
+  request: typeof console.info;
+}
+
 export type PgPool = pg.Pool;
 
 // Copied from import('@types/pg').
