+++ package.json
@@ -1,18 +1,15 @@
 {
-  "name": "@filecoin-station/simple-subnet-api-monorepo",
+  "name": "@filecoin-station/simple-subnet-api",
   "private": true,
   "version": "0.1.0",
   "type": "module",
-  "workspaces": [
-    "api",
-    "db"
-  ],
   "scripts": {
-    "migrate": "npm run migrate -w db",
+    "migrate": "node bin/migrate.js",
+    "start": "node bin/simple-subnet-api.js",
     "lint": "prettier --check . && standard",
     "lint:fix": "prettier --write . && standard --fix .",
     "test:types": "tsc -p .",
-    "test:unit": "npm test --workspaces --if-present",
+    "test:unit": "node --test --test-reporter=spec",
     "test": "npm run lint && npm run test:types && npm run test:unit"
   },
   "repository": {
@@ -25,6 +22,16 @@
     "url": "https://github.com/CheckerNetwork/simple-subnet-api/issues"
   },
   "homepage": "https://github.com/CheckerNetwork/simple-subnet-api#readme",
+  "dependencies": {
+    "@fastify/postgres": "^6.0.2",
+    "@sentry/node": "^9.2.0",
+    "@sentry/profiling-node": "^9.2.0",
+    "debug": "^4.4.0",
+    "fastify": "^5.2.1",
+    "http-assert": "^1.5.0",
+    "pg": "^8.13.3",
+    "postgrator": "^8.0.0"
+  },
   "devDependencies": {
     "@types/pg": "^8.11.11",
     "prettier": "^3.5.2",
